[FIX] Composition API and outsourcing of SideBarFunctions
